Correct upload percents

+++ PangaeaBackend/cores/core.h
@@ -48,7 +48,7 @@ public slots:
 
     void parseInputData(QByteArray data);
     void pushCommandToQueue(QByteArray command, bool finalize = true);
-    void sendWithoutConfirmation(QByteArray data, qint64 dataSizeToSend = -1, qint64 dataSizeTorecieve = -1);
+    void sendWithoutConfirmation(QByteArray data, qint32 dataSizeToSend = -1, qint32 dataSizeTorecieve = -1);
     void processCommands();
 
 private slots:
